add scattered stars to the story title page and table of contents
Extracts DecoStars into a shared component with a dense mode that
adds extra stars. The title and Inhaltsverzeichnis pages use dense
mode for a richer look.
